resume_analyzer_bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pdfplumber==0.10.3
PyPDF2==3.0.1
reportlab==4.0.9
nltk==3.8.1
scikit-learn==1.3.2
openai==1.3.6
google-generativeai==0.3.0
requests==2.31.0
aiohttp==3.9.1
